use a single ManifestModelMock definition
