+++ test/unit/mocks/ManifestModelMock.js
@@ -1,0 +1,11 @@
+class ManifestModelMock {
+    constructor() {
+        this.manifestValue = 0;
+    }
+
+    getValue() {
+        return this.manifestValue;
+    }
+}
+
+export default ManifestModelMock;
